Refactored and add UT
